backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "API for cashback"
authors = [
    {name = "¨jota-dev00-0¨",email = "¨jpar.dev00@gmail.com¨"}
]
license = "MIT"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false